remove unused lcm. add "USE_ROS" flag instead.
fixes breaking things which was introduced when benoit turned off LCM
