--- src/app/ddDrakeVersion.h.in
@@ -1,6 +1,0 @@
-#ifndef __ddDrakeVersion_h
-#define __ddDrakeVersion_h
-
-#cmakedefine DRAKE_OH_FORK
-
-#endif
